requirements.txt
requests
sqlalchemy
pandas
numpy
python-dotenv
matplotlib
scikit-learn